Organize better the classes between files
